Update BanHang lỗi

+++ API/Models/DTO/BanHang/ProductOrder.cs
@@ -7,8 +7,8 @@
         public string? Size { get; set; }
         public string? Img { get; set; }
         public int Quantity { get; set; }
-        public decimal Price { get; set; }
-        public decimal OldPrice { get; set; }
-        public decimal NewPrice => Price;
+        public float Price { get; set; }
+        public float OldPrice { get; set; }
+        public float NewPrice => Price;
     }
 }
